+++ composer/algorithms/algorithm_registry.py
@@ -7,10 +7,9 @@ from composer.algorithms.hparams import (AlibiHparams, AugMixHparams, BlurPoolHp
                                          ColOutHparams, CutMixHparams, CutOutHparams, FactorizeHparams,
                                          GhostBatchNormHparams, LabelSmoothingHparams, LayerFreezingHparams,
                                          MixUpHparams, ProgressiveResizingHparams, RandAugmentHparams, SAMHparams,
-                                         ScaleScheduleHparams, SelectiveBackpropHparams, SqueezeExciteHparams,
-                                         StochasticDepthHparams)
+                                         ScaleScheduleHparams, SelectiveBackpropHparams, SeqLengthWarmupHparams,
+                                         SqueezeExciteHparams, StochasticDepthHparams)
 from composer.algorithms.no_op_model import NoOpModelHparams
-from composer.algorithms.seq_length_warmup import SeqLengthWarmupHparams
 from composer.algorithms.swa.hparams import SWAHparams
 from composer.core.algorithm import Algorithm
